Staff -> Person presentation
Changing the old staff presentation to use person data instead. Avoids requiring an extra file and download from the CDS, just uses spec.
